regular.mk, regular+kde4: rosa-imagewriter use extended
It's been proposed to add an image writing tool to all of
the desktop regular builds (which is reasonable).

rosa-imagewriter depends on qt5 that's currently found in
lxqt, kde4 and kde5 flavours only so let's put it there.

imagewriter (the openSUSE's original one) goes to the rest of
DE builds: xdg-su would fallback to gksu if none of the desktops
got "detected", and that means several more megabytes of crap
(completely unneeded in WM-based images I think).