regular.mk: rework server{,-ovz}
My failure to recognize that it is a "generic" server that
might need some generalization and not an already specialized
one like OpenVZ HN installer had to be fixed up some day;
today is fine.

Basically, let's move package groups ("checkboxes") and alterator
from server-ovz to server, and maybe beef it up a lil' bit more
later; server-ovz is still far from jeos-ovz but the difference
that looks unmergeable is strict sshd control setup that's going
to bite those unsuspecting, so let's leave it for those of us
who are more suspicious of stray ISOs. :)