regular.mk: reworked regular-server a bit
This starts a sort of "server merger" by consuming
both samba4 and hyperv subflavours into checkboxes.

Rationale is pretty clear: environment -- including
hypervisors -- is just an install/run-time variable,
and the set of initial services is another one;
no need to maintain a distinct image for each value
when we've done enough of that to know we can merge.