basealt.mk: disable powertop service
It appears to be the culprit for the "mouse fell asleep"
problem due to not-so-careful use of autosuspend against
hardware that doesn't implement that correctly.

Let the disable-usb-autosuspend hack remain though as
those wishing to enable powertop service might be helped
by that (I've seen a Logitech mouse willing to exhibit
this problem even in its presence though).