syslinux: tweak showopts
This keyword was misused due to the lack of understanding
of its meaning; thanks zerg@ for the hint, let's hide the
kernel boot parameters not expected to be modified by user
under the hood (and vice versa in some cases).

Note that this is just a keyword shift in all cases
but mediacheck where "showopts" is now added.

NB: this keyword is only relevant for gfxboot.