regular.mk: dropped smartcard support from rescue
It's rare enough that more complete images could be used,
and it pulls in polkit with mozjs which is terrible here.

Never been a feature request but rather a TODO item,
and image size is what folks seem to be actually
concerned with.

Reverts: 29ad2393545820fd8835de95c471b937be7a9514