desktop+lxqt*: reshuffle qt4 bits
This is to make desktop+lxqt packages installable on e2k
as there's no qt4 package within this Sisyphus port and
things might just stay this way (it's long unsupported,
looks like porting software to qt5 is more worthwile
an effort).

There's one more removal though: qupzilla. This browser
fits lxqt like a glove *but* it's pulled in already by
mixin/regular-lxqt, and browser feature is available now
to make the selection more flexible.  It's also going to
be renamed on the way to incusion into KDE SC.