regular.mk: drop xsane from mate
The problem being worked around by this is:
anything in the lists that Requires: webclient
results in rekonq being pulled into the image
(after mate-default requires firefox no more).

The proper fix is to force *_PACKAGES, *_LISTS
and *_REGEXP to be processed in a _single_
transaction for each destination so that
early mis-expansion of virtual packages
doesn't occur when _installing_ those.

This commit should be reverted then.

See-also: https://bugzilla.altlinux.org/30806