mipsel-{bfk3,mitx}: Mark sm705 as master-of-seat on tavolga and bfk3
sm750 is seen as framebuffer-only device, so it's not marked with
master-of-seat tag with current udev (which does this mostly to
dri cards). But without master-of-seat device logind does not
report that the seat "CanGraphical", and thus sddm does not
work for it; there might be other problems, too.

This change adds udev rule that marks sm705's fb device as
master-of-seat.