aarch64-rpi4: add rpi4-boot-switch
Switch of boot mode for Raspberry Pi 4
In Alt, it is customary to load the kernel, dtb and initrd
using u-boot and configuration in extlinux.conf.
Firmware Raspberry Pi 4 loads and modifies dtb
according to config.txt configuration.
U-boot doesn't.
The scripts in this package allow you to switch boot modes:
Firmware loads the kernel, dtb, and initrd. U-boot is not used.
Firmware loads dtb. U-boot loads the kernel and initrd.
U-boot loads the kernel, dtb, and initrd.