x11: add e2k-specific bits
xorg-drv-mga2 is MCST Graphics Array driver for 2D part
of videochip found in Elbrus-1C+ APU (also to be a part
of Elbrus-2C3).

xorg-drv-smi is another fork of xorg-drv-siliconmotion
that's absent in p9/sisyphus; *might* be of some interest
to mipsel folks (they've got yet another fork it seems).

Ah, and do not override targets on e2k.