mixin-basealt.mk: avoid things broken/missing on e2k
Wonder if installer-feature-quota-stage2 was a problem
on other arches but having it in LIVE_PACKAGES fails
image build for me:

  livecd-installer-features: Conflicts: installer-common-stage2

and having it INSTALL2_PACKAGES results in another startup
problem (seems like quotas try to start a bit too early);
not that this one is a blocker but spoils the party nevertheless.

NB: relevant package versions are the same.

Also avoid workstation/extras pkglist: it's strongly hardwired,
seems like last-minute implementation :-/