workstation: organize machine-specific groups
Four checkboxes with "Elbrus 801-PC" prefix look a bit too much,
let's put these under a parent one; the counter-argument is that
selecting that parent checkbox is going to provide message telling
that there are conflicts within -- the proper widget would be a
dropbox with the number of seats but we don't do that just yet
in alterator-pkg.

The autologin group goes there too; still a hack that should be
replaced by proper alterator-users multiseat support.