centaurus/xorg: fix arch support
There's no std-def kernel on e2k and mipsel so far, and there are
different kernels for some aarch64 machines as well;  we'll have to
deal with this problem that's been solved in m-p-d the other way
around (@KFLAVOUR@ substitution in pkglists among the rest).

Also add graphics branding part: must be there regardless of platform,
otherwise an arbitrary one gets pulled in, this time it was
"xalt-kworkstation".  This is another case of the same problem,
m-p-d would solve it with @BRANDING@ substitution where m-p does
generation of variable contents but not file-based pkglists to
be put into the image for runtime selection during install process.

Restrict plymouth support to @X86 as well, just as bootsplash
branding component has already been.