alt-server.mk: put the horse before the tug
There was something wrong about that order of additives
and the base configuration; let's make it straight,
putting the foundation first and architecture-dependent
additions on top of it.

Note that double-colon syntax is specifically allowing
for multiple rules adding up their recipes (and prereqs),
but all of those rules must be specified as double-colon
(this was violated in the previous state of alt-server.mk).

See-also: http://www.gnu.org/software/make/manual/make.html#Double_002dColon
Fixes: 33436b6b1bf144c1912197c78d7ca53bc3589342
Fixes: c06f87539dd6470eddead4a0cf8d62b65020d66b