e2k: add /write.sh into ISO images
There's some peculiarity in writing ISO images for e2k
onto USB Flash media: we need to arrange a partition
and ext2 filesystem with its label passed to propagator;
the former part is what's expected, and the latter one
is what this script does.

Thanks to aerodisk.ru guys for catching a brown paper bag
class typo making format() try to affect the image and not
the drive, by the way.