oem: use/deflogin/root
This change sets the default root password to 'altlinux'
for all the images that employ use/oem feature.
This password will be reset by the configuration wizard at
the first boot and cannot be used to login
to ssh even if it's enabled out-of-the-box. So, 'happy
path' user experience of our distros does not change;
but it helps in troubleshooting the cases
when alterator-setup cannot start e.g. due to broken
graphics.

See https://lists.altlinux.org/pipermail/devel-distro/2021-April/002351.html