grub: look for any way to set default menu item
This prepares the ability to write and read
the default menu item from the grubenv file.
If grub.cfg contains a line like '^default=id',
then the read value is not used.
Setting the default menu item from the grubenv file
will be indented inside the 'if' optator and therefore
does not match the '^default' pattern.