lists/x-e2k: drop multiseat media kludges
gvfs had some multiseat support but wasn't quite there
in practice so we had to fallback from media automount
to "your flash drive is available for mounting" which
was much better than an e.g. 6-seater lottery "who wins
to get the drive automounted and presented, and who --
all the rest -- gets an error message in the face".

Just drop these.