syslinux, grub: add lowmem to install2, rescue
For install2, this is needed to work around error 99-cdrom.sh with method:cdrom.
The error shows up with make-initrd-bootchain, but not propagator, as lowmem is
enabled for propagator in this mode.
For rescue, the point is that the latter has become quite large, so by default
it would be better that way.