basealt.mk: generalize mediaplayers
VLC used to suit everyone (32-/64-bit x86) but not anymore;
let's at least refactor the mess it all turned into by now.

The only non-NOP difference obvious to me should be the addition
of a media player to mipsel variant of alt-workstation; a bugfix
to me as I've first spotted the lack of a mediaplayer in e2k one.