lib/profile.mk: add branding-<BRANDING>-release to PACKAGES_REQUIRES_INITROOT
It was a mistake to add branding package to the initialized chroot. There
were no problems until the package had dependencies. Now, when most branding
has a dependence on alt-os-release, chroot initialization broke.
See also ALT bug 41570.
Branding needed to be added to PACKAGES_REQUIRED_INITROOT so that it is
installed with basesystem.