stage2: stage1 modules list update
Use full PATH, because otherwise extra modules with the same
endings are packed.
Instead of listing modules from the same directory, the path
to the directory is used.
The 10-stage2 list has been cleaned up, but still has a lot
of redundancy. Need added some basic lists for all targets.