centaurus/mate: add a layout switcher for lightdm-gtk-greeter
In the installer, you can set a password for the user using Cyrillic.
In greeter, you can switch the layout with hotkeys and enter such a password.
The problem is that the current layout is not visible if the password is hidden.
The layout indicator in greeter solves this problem.