alt-server: add kernel-modules packages in the image
Аdd maximum kernel-modules packages for the kernels installed in the image.
When installing the system, some kernel modules with device drivers
will not be installed.
When the system administrator later installs additional modules with drivers,
they must be available in the local image repository,
because the network repository may already have a new kernel and new modules.

Added at the request of Evgeny Sinelnikov <sin@altlinux.org>