dev: add systemd-settings-disable-kill-user-processes to builder/live/systemd
It is necessary so that the build session with screen is not aborted when the
ssh connection is broken.

But an open firmware opens an avenue for doing a wide variety of things, some of which are not necessarily related to telephony at all. Indeed, some applications may not even be useful, and only serve as an illustration of what can be achieved when running open source software. One such example is running Doom on the modem, because, ya’ know, everything needs to run Doom. Biktor put together a special (pre)release of the firmware which bundles X11, a VNC server and Chocolate Doom for those of you who wish to try this out. Below you’ll find a short video showing Doom running on the PinePhone’s modem.
It is necessary so that the build session with screen is not aborted when the ssh connection is broken.