Revert "alt-server: add password to encrypted rootfs"
This reverts commit 97826c7c34ed3e7131996ae8eca3c5ac23b325f5.

In the make-initrd-2.27.1-alt1 version, it was possible to take
the name of the encrypted partition from /etc/crypttab.
As a result, it is no longer necessary to manually enter a password
for systemd-cryptsetup.

On the other hand, using the installer-feature-password-systemd-cryptsetup-stage3
package in the installer has the disadvantage that the user is required to re-enter
the encrypted partition password during the postinstall step.
Dropping installer-feature-password-systemd-cryptsetup-stage3 would fix this
at the cost of having to enter the password in the initrd.

See also:
https://bugzilla.altlinux.org/42987
https://bugzilla.altlinux.org/43056
https://bugzilla.altlinux.org/43080