Initial feature live-install
This feature complements the live image of the second stage with installation
image-specific settings and scripted hooks.

It is recommended to connect using +live-installer to ensure that a typical set
of installer-related features is included.

Unlike the install2 feature, a separate image of the second stage of 'altinst'
is not built, but the live image is supplemented with installer packages in
order to reduce the total volume of the iso image.

Installation consists of two stages:
1. The live image is unpacked, as in livecd-install
2. Install additional packages BASE_PACKAGES and package groups THE_GROUPS as in install2.

Unlike install2, only those packages that are not in the live image are placed in the main repository. This is how the size of the iso image is reduced.