workstation/the.pkgs: drop installer-feature-desktop-other-fs-stage2
In modern Linux, mounting third-party filesystems on /mnt/mntpoint
is not only pointless, but also harmful.
Such file systems are much more convenient for users to mount from
a file manager. Users simply do not understand where to look for
these sections. What a user has not done is unknown to the user.