live.mk: initial distro/live-blender
It's been started as an answer to this question:
http://www.opennet.ru/opennews/art.shtml?num=57760#91

I *think* that distro/.live-kiosk is where use/live/runapp
should be, not the downstream distro/.live-games, since
the whole mini-feature is really about kiosks.

OTOH blender is somewhat akin to games in terms of being
a purely multimedia app with 3D acceleration required
for proper interactive mode, and that's what .live-games
strives to provide.

Edited: https://lists.altlinux.org/pipermail/devel-distro/2022-September/002888.html