repo: Add main repo when first time run 'live'
Since the config was created in live, it remained after installation.
Now the config will be created at the first start of live and only
if /image/ALTLinux/ is present (it will not be available when
bootloading via http or ftp). After installing live, the package will
be deleted and the config will not be created at the first start.