initrd-bootchain: use method disk for local boot
The installer has been adapted to work normally with method=disk when booted from cd-rom.
Using method=disk allows you to remove the iso label from the burned flash drive, which
in turn allows you to manipulate partitions in gparted. It should also help with programs
of writing that remove the iso label.