lists: Remove most of the riscv64-specific exceptions
riscv64 port becomes steadily becomes bigger, better
and more complete. This commit updates the package
lists to remove exceptions for software that is
now available from sisyphus_riscv64.

Note that no point is made regarding usability or
necessity of this software for particular riscv64
images; and all this software was not fully tested.
This commit is just a statement that these exceptions
are no longer necessary in order to build an image.
The release managers may want to go through the lists
and exclude something again when it comes to building
and testing actual release images.