install2: cleanup udev/rules.d/64-md-raid-assembly.rules
evms fails to stop a RAID if a RAID is removed and a new one
is created in its place. The problem appeared after we added
mdadm to altinst.
Reported by mcpain@
NOTE: The problem is not observed in live. Either something is
removed from altinst, or something is not installed.