e2k, install2, live, live-install, rescue: generate menu when generate profile
This allows us to add menu items to the menu that cannot be specified
based on the external features of the files contained in the image.
For example, the item 'live' of boot in rescue mode.
In the future, it will be possible to remake generators for syslinux
and grub using a similar principle.